+++ src/photon_platform/publish/app.py
@@ -1,9 +1,23 @@
+
 """
-run the main app
+Command-line interface for the publish tool.
 """
-from .publish import Publish
+import click
+from . import publish
+
+@click.group()
+def cli():
+    """A tool to streamline and enhance the Sphinx documentation framework."""
+    pass
+
+@cli.command()
+def build():
+    """Builds the documentation."""
+    publish.build()
+
+@cli.command()
+def test():
+    """Builds and serves the documentation locally."""
+    publish.test()
 
 
-def run() -> None:
-    reply = Publish().run()
-    print(reply)